+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2023-01-26 22:40:23
 
 - Add details to single image view page
-  - Inculde a link back to the group
+  - Include a link back to the group
 - Improve lazy loading by removing the infinite scroll stuff
 
 ## 2023-01-25 23:27:33